+++ bindings/vue/CHANGELOG.md
@@ -25,6 +25,7 @@ CHANGELOG
  * Stop exposing Vue.js component options objects.
  * Obsolete the plugin option `components`.
    * It should not be able to select which components will be registered.
+ * Make `install` register all the components of vue-onsenui.
 
 1.0.0 (2016-09-15)
 ----